Full changes
Fixed an issue with the "from lemur.plugins import lemur_vault" this did not allow the plugin to run.

Changed the OpenSSL package name within the setup.py as the installable package is named pyOpenSSL

Fixed an issue where a dictionary was assinging configuration variables before the flask app context had been initialised, these have now been placed within the appropriate functions and called when needed.

Fixed an issue with the process_sign_options function where it complained that "DNSName was not subsciptable". This was due to options['extensions']['sub_alt_names']['names'] being a dictionary of objects, rather than a dictionary of dictionaries.

Fixed an issue with the process_role_options fuction where it complained that "key_usage is not iterable" This was due to options['extensions']['key_usage'] was a object rather than an dictionary.

Fixed an issue with the create_certificate function where lemur would complain "not enough items to unpack expected 3 got 2" this was due to the function not returning an external_id back to lemur.